Update Parser.java
Commands: Cannot type undesirable commands
like todo without description with trailing
or leading white space.

This breaks the Cupcake's main method.

Let's use trim() and strip() methods to check
if undesirable command is passed in.

Increases usability and convenience for users
as they can pass in undesirable command action
with spaces and it will be handled.