Update TaskList.java, Cupcake.java, Ui.java
Users can add duplicate copies of the
exact same task.

The list will be flooded with the exact same tasks
due to multiple adds. Will be harder for user to
thus properly keep track of their tasks.

Let's add a detectDuplicates method to check if
the user's task description already exists. If it does,
don't re-add it and let user know.