Update save feature
The tasks added and modified are
saved for future use only if bye
command is used.

Perhaps, based on Task 7 instructions
and for better user functionality,
it is better to save the tasks as and when
even if 'bye' command is not used.

Let's add a method to write the tasks to
file in Storage.java and reference it in
Cupcake.java.

Thus, even if user accidnetally closes the
GUI without typing bye, the content is saved.