Update Cupcake.java & cupcake.jar
When user types bye command, the gui
window remains open and just prints back
a bye message.

The window should ideally close in response
to a bye command.

Let's add a system.exit(0) to run method.

This improves usability of the chatbot
as users need not manually close the chatbot
but can do so with a command.